refactor(vm): relocate skia sources
