+++ TotalCrossVM/src/nm/ui/linux/gfx_Graphics_c.h
@@ -13,7 +13,7 @@ void privateScreenChange(int32 w, int32 h)
 
 #include "../../init/tcsdl.h"
 #ifdef USE_SKIA
-#include "../android/skia.h"
+#include "../skia/skia.h"
 #endif
